#217 fix package of todomvc app
