--- examples/todomvc-example/src/main/java/de/saxsys/mvvmfx/example/todomvc/model/TodoItem.java
@@ -1,44 +1,0 @@
-package de.saxsys.mvvmfx.example.todomvc.model;
-
-import javafx.beans.property.BooleanProperty;
-import javafx.beans.property.SimpleBooleanProperty;
-import javafx.beans.property.SimpleStringProperty;
-import javafx.beans.property.StringProperty;
-
-/**
- * @author manuel.mauky
- */
-public class TodoItem {
-	
-	private StringProperty text = new SimpleStringProperty();
-	
-	private BooleanProperty completed = new SimpleBooleanProperty(false);
-	
-	public TodoItem(String text) {
-		this.text.set(text);
-	}
-	
-	public String getText() {
-		return text.get();
-	}
-	
-	public StringProperty textProperty() {
-		return text;
-	}
-	
-	public void setText(String text) {
-		this.text.set(text);
-	}
-	
-	public boolean isCompleted() {
-		return completed.get();
-	}
-	
-	public BooleanProperty completedProperty() {
-		return completed;
-	}
-	
-	public void setCompleted(boolean completed) {
-		this.completed.set(completed);
-	}
-}
