--- examples/todomvc-example/src/main/java/de/saxsys/mvvmfx/example/todomvc/ui/MainView.java
@@ -1,9 +1,0 @@
-package de.saxsys.mvvmfx.example.todomvc.ui;
-
-import de.saxsys.mvvmfx.FxmlView;
-
-/**
- * @author manuel.mauky
- */
-public class MainView implements FxmlView<MainViewModel> {
-}
