--- examples/todomvc-example/src/main/java/de/saxsys/mvvmfx/example/todomvc/ui/MainViewModel.java
@@ -1,9 +1,0 @@
-package de.saxsys.mvvmfx.example.todomvc.ui;
-
-import de.saxsys.mvvmfx.ViewModel;
-
-/**
- * @author manuel.mauky
- */
-public class MainViewModel implements ViewModel {
-}
