+++ examples/todomvc-example/src/main/java/de/saxsys/mvvmfx/examples/todomvc/ui/MainView.java
@@ -1,0 +1,9 @@
+package de.saxsys.mvvmfx.examples.todomvc.ui;
+
+import de.saxsys.mvvmfx.FxmlView;
+
+/**
+ * @author manuel.mauky
+ */
+public class MainView implements FxmlView<MainViewModel> {
+}
