+++ examples/todomvc-example/src/main/java/de/saxsys/mvvmfx/examples/todomvc/ui/MainViewModel.java
@@ -1,0 +1,9 @@
+package de.saxsys.mvvmfx.examples.todomvc.ui;
+
+import de.saxsys.mvvmfx.ViewModel;
+
+/**
+ * @author manuel.mauky
+ */
+public class MainViewModel implements ViewModel {
+}
