--- examples/todomvc-example/src/main/resources/de/saxsys/mvvmfx/example/todomvc/ui/controls/ControlsView.fxml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import javafx.geometry.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<HBox fx:controller="de.saxsys.mvvmfx.example.todomvc.ui.controls.ControlsView" alignment="CENTER" spacing="20.0"
-	  xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
-	<children>
-		<Label fx:id="itemsLeftLabel" text="X items left"/>
-		<HBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" spacing="10.0">
-			<children>
-				<ToggleButton mnemonicParsing="false" onAction="#all" selected="true" text="All">
-					<toggleGroup>
-						<ToggleGroup fx:id="stateGroup"/>
-					</toggleGroup>
-				</ToggleButton>
-				<ToggleButton mnemonicParsing="false" onAction="#active" text="Active" toggleGroup="$stateGroup"/>
-				<ToggleButton mnemonicParsing="false" onAction="#completed" text="Completed" toggleGroup="$stateGroup"/>
-			</children>
-			<padding>
-				<Insets bottom="5.0" left="5.0" right="5.0" top="5.0"/>
-			</padding>
-		</HBox>
-	</children>
-	<padding>
-		<Insets bottom="5.0" left="5.0" right="5.0" top="5.0"/>
-	</padding>
-</HBox>
