--- examples/todomvc-example/src/main/resources/de/saxsys/mvvmfx/example/todomvc/ui/item/ItemOverviewView.fxml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-
-<AnchorPane maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" minHeight="-Infinity"
-			minWidth="-Infinity" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1"
-			fx:controller="de.saxsys.mvvmfx.example.todomvc.ui.item.ItemOverviewView">
-	<children>
-		<ListView fx:id="items" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0"
-				  AnchorPane.topAnchor="0.0"/>
-	</children>
-</AnchorPane>
