--- examples/todomvc-example/src/main/resources/de/saxsys/mvvmfx/example/todomvc/ui/item/ItemView.fxml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.*?>
-<?import de.jensd.fx.glyphs.fontawesome.*?>
-<?import java.net.*?>
-<?import javafx.scene.control.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-<?import java.net.URL?>
-
-<HBox fx:id="root" alignment="CENTER_LEFT" minHeight="-Infinity" minWidth="-Infinity" styleClass="item_root"
-	  xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1"
-	  fx:controller="de.saxsys.mvvmfx.example.todomvc.ui.item.ItemView">
-	<children>
-		<CheckBox fx:id="completed" mnemonicParsing="false"/>
-		<StackPane alignment="CENTER_LEFT" HBox.hgrow="ALWAYS">
-			<children>
-				<HBox fx:id="contentBox" styleClass="content_box">
-					<children>
-						<Label fx:id="contentLabel" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308"
-							   text="Label" HBox.hgrow="ALWAYS"/>
-						<Button fx:id="deleteButton" mnemonicParsing="false" onAction="#delete">
-							<graphic>
-								<FontAwesomeIcon glyphName="CLOSE" size="1.5em" styleClass="close_icon"/>
-							</graphic>
-						</Button>
-					</children>
-				</HBox>
-				<TextField fx:id="contentInput" promptText="What needs to be done?" visible="false"/>
-			</children>
-		</StackPane>
-	</children>
-	<stylesheets>
-		<URL value="@itemview.css"/>
-	</stylesheets>
-</HBox>
