+++ examples/todomvc-example/src/main/resources/de/saxsys/mvvmfx/examples/todomvc/ui/additems/AddItemsView.fxml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<?import java.net.URL?>
+<HBox styleClass="add_item_root" fx:controller="de.saxsys.mvvmfx.examples.todomvc.ui.additems.AddItemsView"
+	  alignment="CENTER_LEFT" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
+	<children>
+		<CheckBox fx:id="selectAll" mnemonicParsing="false"/>
+		<TextField fx:id="addInput" promptText="What needs to be done?" HBox.hgrow="ALWAYS"/>
+	</children>
+	<stylesheets>
+		<URL value="@additems.css"/>
+	</stylesheets>
+</HBox>
