+++ examples/todomvc-example/src/main/resources/de/saxsys/mvvmfx/examples/todomvc/ui/additems/additems.css
